Fix `changes_in_branch` to not rely on merge-bases too much
* Branches may diverge, after all.
* Better reconcile the stored stack information with what was produced from the graph.
