+++ crates/but-workspace/tests/workspace/ref_info/with_workspace_commit/mod.rs
@@ -1,0 +1,1314 @@
+use bstr::ByteSlice;
+use but_testsupport::{visualize_commit_graph, visualize_commit_graph_all};
+use but_workspace::{head_info, ref_info};
+use gitbutler_stack::StackId;
+use pretty_assertions::assert_eq;
+
+#[test]
+fn remote_ahead_fast_forwardable() -> anyhow::Result<()> {
+    let (mut repo, mut meta) = read_only_in_memory_scenario("remote-advanced-ff")?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    * fb27086 (HEAD -> gitbutler/workspace) GitButler Workspace Commit
+    | * 89cc2d3 (origin/A) change in A
+    |/  
+    * d79bba9 (A) new file in A
+    * c166d42 (origin/main, origin/HEAD, main) init-integration
+    ");
+
+    // Needs a branch for workspace implied by a branch with metadata.
+    add_stack(
+        &mut meta,
+        StackId::from_number_for_testing(1),
+        "A",
+        StackState::InWorkspace,
+    );
+    // We can look at a workspace ref directly (via HEAD)
+    let opts = standard_options();
+    let info = head_info(&repo, &*meta, opts)?;
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/A",
+                        remote_tracking_ref_name: "refs/remotes/origin/A",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(d79bba9, "new file in A\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [
+                            RemoteCommit(89cc2d3, "change in A\n",
+                        ],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+
+    let expected_info = info;
+
+    let at = repo.find_reference("refs/heads/A")?;
+    let info = ref_info(at, &*meta, opts)?;
+    assert_eq!(
+        info, expected_info,
+        "Information doesn't change just because the starting point is different"
+    );
+
+    // Remove remote configuration to have it deduce the remote.
+    // This is from the times when GB wouldn't set up Git remotes, so it's for backward compatibility.
+    repo.config_snapshot_mut()
+        .remove_section("branch", info.stacks[0].name().unwrap().shorten().as_bstr());
+
+    let at = repo.find_reference("refs/heads/A")?;
+    let info = ref_info(at, &*meta, opts)?;
+    assert_eq!(
+        info, expected_info,
+        "Information doesn't change, the remote is inferred"
+    );
+    Ok(())
+}
+
+#[test]
+fn target_ahead_remote_rewritten() -> anyhow::Result<()> {
+    let (repo, mut meta) = read_only_in_memory_scenario("target-ahead-remote-rewritten")?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    * 03d2336 (HEAD -> gitbutler/workspace) GitButler Workspace Commit
+    * d5d3a92 (A) unique local tip
+    * 6ffd040 shared by name
+    * 4cd56ab unique local
+    | * f4ed16f (origin/main) target ahead
+    | | * 50d31c8 (origin/A) unique remote
+    | | * a9954f1 shared by name
+    | |/  
+    |/|   
+    * | 872c22f shared local/remote
+    |/  
+    * c166d42 (origin/main, origin/HEAD, main) init-integration
+    ");
+
+    add_stack(
+        &mut meta,
+        StackId::from_number_for_testing(1),
+        "A",
+        StackState::InWorkspace,
+    );
+    let opts = standard_options();
+    let info = ref_info(repo.find_reference("A")?, &*meta, opts)?;
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/A",
+                        remote_tracking_ref_name: "refs/remotes/origin/A",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(d5d3a92, "unique local tip\n", local),
+                            LocalCommit(6ffd040, "shared by name\n", local/remote(similarity)),
+                            LocalCommit(4cd56ab, "unique local\n", local),
+                            LocalCommit(872c22f, "shared local/remote\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [
+                            RemoteCommit(50d31c8, "unique remote\n",
+                        ],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    Ok(())
+}
+
+#[test]
+fn single_commit_but_two_branches_one_in_ws_commit() -> anyhow::Result<()> {
+    let (repo, mut meta) =
+        read_only_in_memory_scenario("two-branches-one-advanced-one-parent-ws-commit")?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    *   7f3248e (HEAD -> gitbutler/workspace) GitButler Workspace Commit
+    |\  
+    | * cbc6713 (advanced-lane) change
+    * | 93d7eac (advanced-lane-2) change 2
+    |/  
+    * fafd9d0 (origin/main, main, lane) init
+    ");
+
+    for (idx, name) in ["lane", "advanced-lane-2", "advanced-lane"]
+        .into_iter()
+        .enumerate()
+    {
+        add_stack(
+            &mut meta,
+            StackId::from_number_for_testing(idx as u128),
+            name,
+            StackState::InWorkspace,
+        );
+    }
+
+    let opts = standard_options();
+    let info = head_info(&repo, &*meta, opts)?;
+    // The difficulty here is that there is no merge-parent for the newly created stack, and that
+    // empty stacks rest on the merge-base which naturally is hidden during traversal.
+    // Also, according to
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/advanced-lane-2",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(93d7eac, "change 2\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/advanced-lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(cbc6713, "change\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    Ok(())
+}
+
+#[test]
+fn single_commit_but_two_branches_one_in_ws_commit_with_virtual_segments() -> anyhow::Result<()> {
+    let (repo, mut meta) =
+        read_only_in_memory_scenario("multiple-dependent-branches-per-stack-without-commit")?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    * cbc6713 (HEAD -> gitbutler/workspace, lane) change
+    * fafd9d0 (origin/main, main, lane-segment-02, lane-segment-01, lane-2-segment-02, lane-2-segment-01, lane-2) init
+    ");
+
+    // Follow the natural order, lane first.
+    add_stack_with_segments(
+        &mut meta,
+        StackId::from_number_for_testing(0),
+        "lane",
+        StackState::InWorkspace,
+        &["lane-segment-01", "lane-segment-02"],
+    );
+    add_stack_with_segments(
+        &mut meta,
+        StackId::from_number_for_testing(1),
+        "lane-2",
+        StackState::InWorkspace,
+        &["lane-2-segment-01", "lane-2-segment-02"],
+    );
+
+    // The stacks should come out just like defined above, "lane" and then "lane2" with all the right segments.
+    let opts = standard_options();
+    let info = ref_info(repo.find_reference("lane")?, &*meta, opts)?;
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(cbc6713, "change\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/lane-2-segment-02",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+
+    // Natural order here is `lane` first, but we say we want `lane-2` first
+    add_stack_with_segments(
+        &mut meta,
+        StackId::from_number_for_testing(1),
+        "lane-2",
+        StackState::InWorkspace,
+        &["lane-2-segment-01", "lane-2-segment-02"],
+    );
+    add_stack_with_segments(
+        &mut meta,
+        StackId::from_number_for_testing(0),
+        "lane",
+        StackState::InWorkspace,
+        &["lane-segment-01", "lane-segment-02"],
+    );
+
+    let opts = standard_options();
+    let info = ref_info(repo.find_reference("lane")?, &*meta, opts)?;
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(cbc6713, "change\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/lane-2-segment-02",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    Ok(())
+}
+
+#[test]
+fn single_commit_but_two_branches_both_in_ws_commit() -> anyhow::Result<()> {
+    let (repo, mut meta) =
+        read_only_in_memory_scenario("two-branches-one-advanced-two-parent-ws-commit")?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    *   335d6f2 (HEAD -> gitbutler/workspace) GitButler Workspace Commit
+    |\
+    | * cbc6713 (advanced-lane) change
+    |/
+    * fafd9d0 (origin/main, main, lane) init
+    ");
+
+    for (idx, name) in ["advanced-lane", "lane"].into_iter().enumerate() {
+        add_stack(
+            &mut meta,
+            StackId::from_number_for_testing(idx as u128),
+            name,
+            StackState::InWorkspace,
+        );
+    }
+
+    let opts = standard_options();
+    let info = head_info(&repo, &*meta, opts)?;
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/advanced-lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(cbc6713, "change\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    Ok(())
+}
+
+#[test]
+fn single_commit_but_two_branches_stack_on_top_of_ws_commit() -> anyhow::Result<()> {
+    let (repo, mut meta) =
+        read_only_in_memory_scenario("two-branches-one-advanced-ws-commit-on-top-of-stack")?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    * cbc6713 (HEAD -> gitbutler/workspace, advanced-lane) change
+    * fafd9d0 (origin/main, main, lane) init
+    ");
+
+    for (idx, name) in ["advanced-lane", "lane"].into_iter().enumerate() {
+        add_stack(
+            &mut meta,
+            StackId::from_number_for_testing(idx as u128),
+            name,
+            StackState::InWorkspace,
+        );
+    }
+
+    let opts = standard_options();
+    let err = head_info(&repo, &*meta, opts).unwrap_err();
+    assert_eq!(
+        err.to_string(),
+        "Workspace reference gitbutler/workspace didn't point to a managed commit anymore",
+        "we simply don't know what to do here yet"
+    );
+
+    let info = ref_info(repo.find_reference("advanced-lane")?, &*meta, opts).unwrap();
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/advanced-lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(cbc6713, "change\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+            Stack {
+                base: Some(
+                    Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+
+    Ok(())
+}
+
+#[test]
+fn two_branches_one_advanced_two_parent_ws_commit_diverged_remote_tracking_branch()
+-> anyhow::Result<()> {
+    let (repo, mut meta) = read_only_in_memory_scenario(
+        "two-branches-one-advanced-two-parent-ws-commit-diverged-ttb",
+    )?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    *   873d056 (HEAD -> gitbutler/workspace) GitButler Workspace Commit
+    |\  
+    * | cbc6713 (advanced-lane) change
+    |/  
+    * fafd9d0 (main, lane) init
+    * da83717 (origin/main) disjoint remote target
+    ");
+
+    for (idx, name) in ["advanced-lane", "lane"].into_iter().enumerate() {
+        add_stack(
+            &mut meta,
+            StackId::from_number_for_testing(idx as u128),
+            name,
+            StackState::InWorkspace,
+        );
+    }
+
+    let opts = standard_options();
+    let info = head_info(&repo, &*meta, opts)?;
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: None,
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/advanced-lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(cbc6713, "change\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+            Stack {
+                base: None,
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/lane",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    Ok(())
+}
+
+#[test]
+fn disjoint() -> anyhow::Result<()> {
+    let (repo, mut meta) = read_only_in_memory_scenario("disjoint")?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    * 32791d2 (HEAD -> disjoint) disjoint init
+    * fafd9d0 (origin/main, main) init
+    ");
+
+    add_stack(
+        &mut meta,
+        StackId::from_number_for_testing(1),
+        "disjoint",
+        StackState::InWorkspace,
+    );
+
+    let opts = standard_options();
+    let info = head_info(&repo, &*meta, opts)?;
+
+    // We see the commit in the branch as there is no base to hide it.
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: None,
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/disjoint",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "OutsideOfWorkspace",
+                        commits_unique_from_tip: [
+                            LocalCommit(32791d2, "disjoint init\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    Ok(())
+}
+
+#[test]
+fn multiple_branches_with_shared_segment() -> anyhow::Result<()> {
+    let (repo, mut meta) = read_only_in_memory_scenario("multiple-stacks-with-shared-segment")?;
+    insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    *   820f2b3 (HEAD -> gitbutler/workspace) GitButler Workspace Commit
+    |\  
+    | * 4e5484a (B-on-A) add new file in B-on-A
+    * | 5f37dbf (C-on-A) add new file in C-on-A
+    |/  
+    | * 89cc2d3 (origin/A) change in A
+    |/  
+    * d79bba9 (A) new file in A
+    * c166d42 (origin/main, origin/HEAD, main) init-integration
+    ");
+
+    add_stack(
+        &mut meta,
+        StackId::from_number_for_testing(1),
+        "C-on-A",
+        StackState::InWorkspace,
+    );
+
+    let opts = standard_options();
+    let info = head_info(&repo, &*meta, opts)?;
+
+    // The shared "A" segment is only used in the first stack, despite reachable from both.
+    // The first stack which reaches the shared segment claims it.
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/C-on-A",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(5f37dbf, "add new file in C-on-A\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                    StackSegment {
+                        ref_name: "refs/heads/A",
+                        remote_tracking_ref_name: "refs/remotes/origin/A",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(d79bba9, "new file in A\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [
+                            RemoteCommit(89cc2d3, "change in A\n",
+                        ],
+                        metadata: None,
+                    },
+                ],
+                stash_status: None,
+            },
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/B-on-A",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(4e5484a, "add new file in B-on-A\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: None,
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+
+    let c_info = ref_info(repo.find_reference("C-on-A")?, &*meta, opts)?;
+
+    // A partial workspace is provided, but the entire workspace is known.
+    insta::assert_debug_snapshot!(c_info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/C-on-A",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(5f37dbf, "add new file in C-on-A\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                    StackSegment {
+                        ref_name: "refs/heads/A",
+                        remote_tracking_ref_name: "refs/remotes/origin/A",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(d79bba9, "new file in A\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [
+                            RemoteCommit(89cc2d3, "change in A\n",
+                        ],
+                        metadata: None,
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    assert_eq!(info.stacks[0], c_info.stacks[0]);
+
+    let b_info = ref_info(repo.find_reference("B-on-A")?, &*meta, opts)?;
+
+    // It's like the stack is part of the workspace, so "A" is only used once.
+    insta::assert_debug_snapshot!(b_info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/B-on-A",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(4e5484a, "add new file in B-on-A\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: None,
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+
+    assert_eq!(
+        info.stacks[1].segments, b_info.stacks[0].segments,
+        "It's like the stack is part of the workspace, so 'A' is only used once."
+    );
+
+    let a_info = ref_info(repo.find_reference("A")?, &*meta, opts)?;
+
+    // We can also show segments that are part of the stack (like homing in on them), as long as they are in a workspace.
+    insta::assert_debug_snapshot!(a_info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/A",
+                        remote_tracking_ref_name: "refs/remotes/origin/A",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [
+                            LocalCommit(d79bba9, "new file in A\n", local),
+                        ],
+                        commits_unique_in_remote_tracking_branch: [
+                            RemoteCommit(89cc2d3, "change in A\n",
+                        ],
+                        metadata: None,
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    Ok(())
+}
+
+#[test]
+fn empty_workspace_with_branch_below() -> anyhow::Result<()> {
+    let (repo, mut meta) = read_only_in_memory_scenario("empty-workspace-with-branch-below")?;
+    insta::assert_snapshot!(visualize_commit_graph(&repo, "HEAD")?, @r"
+    * c7276fa (HEAD -> gitbutler/workspace) GitButler Workspace Commit
+    * c166d42 (origin/main, origin/HEAD, unrelated, main) init-integration
+    ");
+
+    add_stack(
+        &mut meta,
+        StackId::from_number_for_testing(1),
+        "unrelated",
+        StackState::InWorkspace,
+    );
+
+    let opts = standard_options();
+    let info = head_info(&repo, &*meta, opts)?;
+    // Active branches we should see, but only "unrelated",
+    // not any other branch that happens to point at that commit.
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/unrelated",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "ReachableFromWorkspaceCommit",
+                        commits_unique_from_tip: [],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+
+    // Change the stack to be inactive, so it's not considered to be part of the workspace.
+    add_stack(
+        &mut meta,
+        StackId::from_number_for_testing(1),
+        "unrelated",
+        StackState::Inactive,
+    );
+
+    let info = head_info(&repo, &*meta, opts)?;
+    // Now there should be no stack, it's an empty workspace.
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+
+    // But if it's requested directly, we should see it nonetheless.
+    let info = ref_info(repo.find_reference("unrelated")?, &*meta, opts)?;
+    insta::assert_debug_snapshot!(info, @r#"
+    RefInfo {
+        workspace_ref_name: Some(
+            FullName(
+                "refs/heads/gitbutler/workspace",
+            ),
+        ),
+        stacks: [
+            Stack {
+                base: Some(
+                    Sha1(c166d42d4ef2e5e742d33554d03805cfb0b24d11),
+                ),
+                segments: [
+                    StackSegment {
+                        ref_name: "refs/heads/unrelated",
+                        remote_tracking_ref_name: "None",
+                        ref_location: "OutsideOfWorkspace",
+                        commits_unique_from_tip: [],
+                        commits_unique_in_remote_tracking_branch: [],
+                        metadata: Some(
+                            Branch {
+                                ref_info: RefInfo { created_at: None, updated_at: "1970-01-01 00:00:00 +0000" },
+                                description: None,
+                                review: Review { pull_request: None, review_id: None },
+                            },
+                        ),
+                    },
+                ],
+                stash_status: None,
+            },
+        ],
+        target_ref: Some(
+            FullName(
+                "refs/remotes/origin/main",
+            ),
+        ),
+    }
+    "#);
+    Ok(())
+}
+
+mod legacy;
+
+mod branch_details {
+    use crate::ref_info::with_workspace_commit::read_only_in_memory_scenario;
+    use but_testsupport::visualize_commit_graph_all;
+    use but_workspace::branch_details_v3;
+
+    #[test]
+    fn disjoint() -> anyhow::Result<()> {
+        let (repo, meta) = read_only_in_memory_scenario("disjoint")?;
+        insta::assert_snapshot!(visualize_commit_graph_all(&repo)?, @r"
+    * 32791d2 (HEAD -> disjoint) disjoint init
+    * fafd9d0 (origin/main, main) init
+    ");
+
+        let actual = branch_details_v3(&repo, "refs/heads/disjoint".try_into()?, &*meta)?;
+        insta::assert_debug_snapshot!(actual, @r#"
+        BranchDetails {
+            name: "refs/heads/disjoint",
+            remote_tracking_branch: None,
+            description: None,
+            pr_number: None,
+            review_id: None,
+            tip: Sha1(32791d22e276ec0ed87d14f906321137356bc6d6),
+            base_commit: Sha1(32791d22e276ec0ed87d14f906321137356bc6d6),
+            push_status: CompletelyUnpushed,
+            last_updated_at: None,
+            authors: [
+                author <author@example.com>,
+                committer <committer@example.com>,
+            ],
+            is_conflicted: false,
+            commits: [
+                Commit(32791d2, "disjoint init", local/remote(identity)),
+            ],
+            upstream_commits: [],
+            is_remote_head: false,
+        }
+        "#);
+
+        let actual = branch_details_v3(&repo, "refs/heads/main".try_into()?, &*meta)?;
+        insta::assert_debug_snapshot!(actual, @r#"
+        BranchDetails {
+            name: "refs/heads/main",
+            remote_tracking_branch: None,
+            description: None,
+            pr_number: None,
+            review_id: None,
+            tip: Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+            base_commit: Sha1(fafd9d08a839d99db60b222cd58e2e0bfaf1f7b2),
+            push_status: CompletelyUnpushed,
+            last_updated_at: None,
+            authors: [],
+            is_conflicted: false,
+            commits: [],
+            upstream_commits: [],
+            is_remote_head: false,
+        }
+        "#);
+
+        Ok(())
+    }
+}
+mod utils {
+    use crate::ref_info::utils::named_read_only_in_memory_scenario;
+    use but_workspace::VirtualBranchesTomlMetadata;
+    use gitbutler_stack::StackId;
+
+    pub fn read_only_in_memory_scenario(
+        name: &str,
+    ) -> anyhow::Result<(
+        gix::Repository,
+        std::mem::ManuallyDrop<VirtualBranchesTomlMetadata>,
+    )> {
+        let (repo, mut meta) =
+            named_read_only_in_memory_scenario("with-remotes-and-workspace", name)?;
+        let vb = meta.data_mut();
+        vb.default_target = Some(gitbutler_stack::Target {
+            // For simplicity, we stick to the defaults.
+            branch: gitbutler_reference::RemoteRefname::new("origin", "main"),
+            // Not required
+            remote_url: "should not be needed and when it is extract it from `repo`".to_string(),
+            sha: git2::Oid::zero(),
+            push_remote_name: None,
+        });
+        Ok((repo, meta))
+    }
+
+    pub enum StackState {
+        InWorkspace,
+        Inactive,
+    }
+
+    // Add parameters as needed.
+    pub fn add_stack(
+        meta: &mut VirtualBranchesTomlMetadata,
+        stack_id: StackId,
+        stack_name: &str,
+        state: StackState,
+    ) -> StackId {
+        add_stack_with_segments(meta, stack_id, stack_name, state, &[])
+    }
+
+    // Add parameters as needed.
+    pub fn add_stack_with_segments(
+        meta: &mut VirtualBranchesTomlMetadata,
+        stack_id: StackId,
+        stack_name: &str,
+        state: StackState,
+        segments: &[&str],
+    ) -> StackId {
+        let mut stack = gitbutler_stack::Stack::new_with_just_heads(
+            segments
+                .iter()
+                .rev()
+                .map(|stack_name| {
+                    gitbutler_stack::StackBranch::new_with_zero_head(
+                        (*stack_name).into(),
+                        None,
+                        None,
+                        None,
+                        false,
+                    )
+                })
+                .chain(std::iter::once(
+                    gitbutler_stack::StackBranch::new_with_zero_head(
+                        stack_name.into(),
+                        None,
+                        None,
+                        None,
+                        false,
+                    ),
+                ))
+                .collect(),
+            0,
+            meta.data().branches.len(),
+            match state {
+                StackState::InWorkspace => true,
+                StackState::Inactive => false,
+            },
+        );
+        stack.id = stack_id;
+        meta.data_mut().branches.insert(stack_id, stack);
+        meta.data();
+        stack_id
+    }
+}
+use crate::ref_info::utils::standard_options;
+use crate::ref_info::with_workspace_commit::utils::{StackState, add_stack_with_segments};
+use utils::add_stack;
+pub use utils::read_only_in_memory_scenario;
